requirements.txt
torch>=2.1.0
lightning @ git+https://github.com/Lightning-AI/lightning@6cbe9ceb560d798892bdae9186291acf9bf5d2e3
jsonargparse[signatures]
bitsandbytes==0.41.0   
scipy                   
sentencepiece           
tokenizers              
datasets                
zstandard               
pandas                 
pyarrow             
tensorboard         
requests==2.31.0
bs4==4.12.2
html2text==2020.1.16
openai==1.3.8
python-dotenv==1.0.0
